requirements.txt
pyserial
keyboard
numpy
scipy